refactor: extract service actions menu and tabs list components
