+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/_dashboard/services/-services/service-tabs-list.tsx
@@ -1,0 +1,12 @@
+import { TabsList, TabsTrigger } from "@packages/ui/components/tabs";
+
+export function ServiceTabsList() {
+   return (
+      <TabsList>
+         <TabsTrigger value="precos">Preços</TabsTrigger>
+         <TabsTrigger value="beneficios">Benefícios</TabsTrigger>
+         <TabsTrigger value="assinantes">Assinantes</TabsTrigger>
+         <TabsTrigger value="overview">Overview</TabsTrigger>
+      </TabsList>
+   );
+}
